+++ Zipadoo/Zipadoo/.swiftlint.yml
@@ -9,6 +9,11 @@ line_length:
 #indentation:
 #  tabs: true
 
+# 코드 줄 길이 설정
+type_body_length:
+  warning: 400
+  error: 500
+  
 function_body_length:
   enabled: false
